Fix major bug breaking every custom command.
Testing using a delay command was useless.  I completely missed the fact that the command didn't execute because it produced no output. 

Anyway this PR changes the command to use `start` meaning that the sub process is executed in another thread. This may fix the issue of the sensor turning off in hass.io.